Fix buffer head shift because of advance len check
When the message size is very large, I encounter the issue that the
message is not correctly decoded.
The bug was from using get_u32 which will advance the buffer there fore
the next read of buffer for message len give the wrong length of stream.
Change to use peek method solve the problem.

Thanks chatGPT.